pyproject.toml
[tool.poetry]
name = "mazzeh-chin"
version = "0.1.0"
description = ""
authors = ["mr.khavari <mr.khavari.n@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.2"
pydantic-settings = "^2.6.0"
neomodel = "^5.3.3"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
passlib = "^1.7.4"
pyjwt = "^2.9.0"
python-multipart = "^0.0.12"


[tool.poetry.group.dev.dependencies]
black = {version = "^24.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
